Updating for new APIs

+++ Bttendance/controller/CourseAttendView.h
@@ -19,13 +19,12 @@
 
     CourseInfoCell *currentcell;
 
-    int sid;
+    NSInteger sid;
     NSString *sname;
-
 }
 @property(weak, nonatomic) IBOutlet UITableView *tableview;
 
-@property(nonatomic) int sid;
+@property(nonatomic) NSInteger sid;
 @property(nonatomic) NSString *sname;
 
 - (IBAction)check_button_action:(id)sender;
